==> notebooks/etl/01_etl_appointment.ipynb <==
# ETL - Atendimento


from pathlib import Path
from typing import Dict, List, Tuple
import pandas as pd

root_dir = Path('../../data')
raw_dir = Path('raw/20250101_carga_inicial_ana_roccio')
gold_dir = Path('gold')
str_atend = 'TAB_ATENDIMENTO.csv'
str_appoint = 'TAB_ATENDIMENTO_ANALISE.csv'
str_ubs = 'TAB_UNIDADE_SAUDE.csv'
str_target = 'appointments.parquet'

df_ubs = pd.read_csv(root_dir / raw_dir / str_ubs, low_memory=False)
df_ubs.columns = df_ubs.columns.str.lower()
df_ubs

df_atend = pd.read_csv(root_dir / raw_dir / str_atend, low_memory=False)
df_atend.columns = df_atend.columns.str.lower()
df_atend = df_atend.drop(columns=['cod_paciente', 'data_atendimento', 'especialidade', 'periodo_extracao'])
df_atend

df_atend = pd.merge(df_atend, df_ubs, on='cod_unidade_saude')
df_atend

df_appoint = pd.read_csv(root_dir / raw_dir / str_appoint, low_memory=False)
df_appoint.columns = df_appoint.columns.str.lower()
df_appoint = df_appoint.drop(columns=['periodo_extracao'])

df_appoint.data_atendimento = pd.to_datetime(df_appoint.data_atendimento)
df_appoint['e_diag_infeccioso'] = df_appoint['e_diag_infeccioso'].astype(bool)
df_appoint = df_appoint.fillna('Não Especificado')
df_appoint

df_appoint['ano']       = df_appoint['data_atendimento'].dt.year
df_appoint['mes']       = df_appoint['data_atendimento'].dt.month
df_appoint['nome_mes']  = df_appoint['data_atendimento'].dt.month_name()
df_appoint['ano_mes']   = df_appoint['data_atendimento'].dt.to_period('M').astype(str)
df_appoint

df_appoint = pd.merge(df_appoint, df_atend, on='cod_atendimento')
df_appoint

dftmp = df_appoint.groupby('cod_atendimento')['cod_cid_ciap'].nunique().reset_index(name='n_cid_ciap')
dftmp

df_appoint = pd.merge(df_appoint, dftmp, on='cod_atendimento')
df_appoint

def calcular_faixa_etaria(idade):
    if pd.isna(idade):
        return 'Não informado'
    
    idade = int(idade)
    if idade < 1:
        return '0-1 ano'
    elif idade < 12:
        return '1-11 anos'
    elif idade < 18:
        return '12-17 anos'
    elif idade < 60:
        return '18-59 anos'
    else:
        return '60+ anos'
    
df_appoint['faixa_etaria'] = df_appoint['idade'].apply(calcular_faixa_etaria)
df_appoint

full_path = root_dir / gold_dir
full_path.mkdir(parents=True, exist_ok=True)
df_appoint.to_parquet(full_path / str_target, index=False)
