==> notebooks/etl/02_etl_prescription.ipynb <==
# ETL - Prescricao


from pathlib import Path
from typing import Dict, List, Tuple
import pandas as pd

root_dir = Path('../../data')
raw_dir = Path('raw/20250101_carga_inicial_ana_roccio')
gold_dir = Path('gold')
str_med = 'TAB_MEDICAMENTO.csv'
str_pred = 'TAB_MED_PRESCRITO.csv'
str_fname = 'TAB_MEDPRESCRITO_ANALISE.csv'
str_appoint = 'appointments.parquet'
str_target = 'full_data.parquet'

df_presc = pd.read_csv(root_dir / raw_dir / str_fname, low_memory=False)
df_presc.columns = df_presc.columns.str.lower()
df_presc

df_presc['e_antibiotico'] = df_presc['e_antibiotico'].astype(bool)
df_presc = df_presc.fillna('Não Especificado')
df_presc

dftmp = df_presc.groupby('cod_atendimento')['cod_medicamento'].nunique().reset_index(name='n_medicacao')
dftmp

df_presc = pd.merge(df_presc, dftmp, on='cod_atendimento')
df_presc

df_appoint = pd.read_parquet(root_dir / gold_dir / str_appoint)
df_appoint

df = pd.merge(df_appoint, df_presc, on='cod_atendimento')
df

# Flag de prescrição inadequada
df['e_presc_inadequada'] = ((df['e_antibiotico'] == True) & (df['e_diag_infeccioso'] == False) | 
                            (df['e_antibiotico'] == False) & (df['e_diag_infeccioso'] == True))
df

full_path = root_dir / gold_dir
full_path.mkdir(parents=True, exist_ok=True)
df.to_parquet(full_path / str_target, index=False)
